bug: Fix issue with command data mapping
When the response containing command data was mapped it was wrapped in a "commandJson" object and the properties of that were not being accessed correctly. This change unwraps the command data from the "commandJson" object. Just reading the node within in "commandJson" would return a String type node which is then mapped to a root node so the properties can be accessed.