refactor: refactor event handling
This change removed the old event handling using dispatchers, registries, and module initializers as well as added a new parent class EventListener as the parent of all event listeners. All events in modules now implement EventSubscriber all services have been moved to the app module and are declared in the io.datcord.event.EventSubscriber file with their fully qualified names. Events are now loaded in the EventDispatcher this includes events found in modules.

Note - the jira command/feature is not fully developed and is only included in this push to get the major changes to event handling ASAP

Note** - A lot of the new code related to event handling still needs to be documented

- Deleted event interaction dispatcher and registry classes.
- Removed EventModuleInitializer interface.
- Deleted command module for handling ticket events.
- Added new module definitions for jira interaction handling.
- Updated app's pom.xml to include the jira module as a dependency.