feat: Upgrade Jira assignee functionality and add Discord user mapping
The JiraAssigneeWebhookMessageReceivedEventListener class has been updated to perform a GET request to the Jira association endpoint when an assignee field is found in a webhook message. A ResponseBuilder class was also added to make a POST request for creating an association and map string or long fields. Assignment mapping from Jira user to Discord user has also been added. A new jiraAssociationDto class was created for mapping Jira associations, and the deprecated ReadyEventListener class was marked for removal. Further work will improve the reusability of these code updates.